Update VSIX
Update VSIX to support future VS.  Fix the gitignore file to include
nuget packages in the git.  Include all the needed nuget packages so
it's independent of the packages installed by Visual Studio.  Link to js
files in the sample project so we don't have to copy it to vsix every
time there is a change.