Allow VSIX file created without any nuget package
Allow VSIX file created without any nuget package included in the VSIX
file.  We rely on Nuget 2.7 auto restore package during build.  However,
we have a risk here that if one of the dependent package is withdrawn
from nuget.org, then we'll break VSIX creation.  Also offline creation
will not work.  So this is just a demo for now.